+++ packages/template-chakra-storefront/src/pages/account/wishlist/partials/wishlist-primary-action.test.js
@@ -11,20 +11,32 @@ import {renderWithProviders} from '../../../../utils/test-utils'
 import WishlistPrimaryAction from './wishlist-primary-action'
 import {screen, waitFor, act} from '@testing-library/react'
 import PropTypes from 'prop-types'
-import {rest} from 'msw'
 import {basketWithProductSet} from '../../../product-detail/index.mock'
 import {mockProductBundle} from '../../../../../mocks/product-bundle'
+import {prependHandlersToServer} from '../../../../../jest-setup'
+import Toaster, {toaster} from '../../../../components/toaster'
 
 const MockedComponent = ({variant}) => {
     return (
         <ItemVariantProvider variant={variant}>
             <WishlistPrimaryAction />
+            <Toaster toaster={toaster} />
         </ItemVariantProvider>
     )
 }
 MockedComponent.propTypes = {
     variant: PropTypes.object
 }
+
+// Mock the useShopperBasketsMutationHelper hook
+const mockAddItemToNewOrExistingBasket = jest.fn()
+
+jest.mock('@salesforce/commerce-sdk-react', () => ({
+    ...jest.requireActual('@salesforce/commerce-sdk-react'),
+    useShopperBasketsMutationHelper: jest.fn(() => ({
+        addItemToNewOrExistingBasket: mockAddItemToNewOrExistingBasket
+    }))
+}))
 
 jest.mock('../../../../hooks/use-current-basket', () => {
     return {
@@ -40,22 +52,37 @@ jest.mock('../../../../hooks/use-current-basket', () => {
 beforeEach(() => {
     jest.resetModules()
 
-    global.server.use(
-        // For adding items to basket
-        rest.post('*/baskets/:basketId/items', (req, res, ctx) => {
-            return res(ctx.json(basketWithProductSet))
-        })
-    )
+    prependHandlersToServer([
+        {
+            path: '*/baskets/:basketId/items',
+            method: 'post',
+            res: () => basketWithProductSet
+        }
+    ])
 })
 
-// TODO: unskip this test when we have a more proper mock for the add set to cart API call
-test.skip('the Add To Cart button', async () => {
+afterEach(() => {
+    jest.restoreAllMocks()
+})
+
+test('the Add To Cart button', async () => {
+    // 701642884934M
     const variant = mockWishListDetails.data[3]
 
     const {user} = renderWithProviders(<MockedComponent variant={variant} />)
 
     const addToCartButton = await screen.findByRole('button', {
         name: new RegExp(`Add ${variant.name} to cart`, 'i')
+    })
+
+    mockAddItemToNewOrExistingBasket.mockResolvedValue({
+        basketId: 'basket_id',
+        productItems: [
+            {
+                productId: variant.id,
+                productName: variant.name
+            }
+        ]
     })
     await act(async () => {
         await user.click(addToCartButton)
@@ -66,12 +93,25 @@ test.skip('the Add To Cart button', async () => {
     })
 })
 
-// TODO: unskip this test when we have a more proper mock for the add set to cart API call
-test.skip('the Add Set To Cart button', async () => {
+test('the Add Set To Cart button', async () => {
     const productSetWithoutVariants = mockWishListDetails.data[1]
     const {user} = renderWithProviders(<MockedComponent variant={productSetWithoutVariants} />)
     const addSetToCartButton = await screen.findByRole('button', {
         name: new RegExp(`Add ${productSetWithoutVariants.name} set to cart`, 'i')
+    })
+    // set product children being added to cart
+    mockAddItemToNewOrExistingBasket.mockResolvedValue({
+        basketId: 'basket_id',
+        productItems: [
+            {
+                productId: productSetWithoutVariants.setProducts[0].id,
+                productName: productSetWithoutVariants.setProducts[0].name
+            },
+            {
+                productId: productSetWithoutVariants.setProducts[0].id,
+                productName: productSetWithoutVariants.setProducts[0].name
+            }
+        ]
     })
     await act(async () => {
         await user.click(addSetToCartButton)